requirements.txt
# Production requirements for AWS EC2 deployment
# Compatible with Python 3.8+ and Ubuntu 22.04 LTS

# Core Flask and web framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7

# Telegram Bot API
python-telegram-bot==13.15

# Database
SQLAlchemy==1.4.49
psycopg2-binary==2.9.7  # PostgreSQL adapter

# HTTP and networking
requests==2.31.0
urllib3==2.0.7
certifi==2023.7.22

# Cryptography and security
cryptography==41.0.4
cffi==1.15.1
pycparser==2.21

# Utilities
six==1.16.0
chardet==5.2.0
idna==3.4
future==0.18.3
dnspython==2.4.2
python-http-client==3.3.7

# Production WSGI server
gunicorn==21.2.0

# Environment management
python-dotenv==1.0.0

# AWS SDK (optional, for future AWS integrations)
boto3==1.28.57
botocore==1.31.57

# Monitoring and logging
psutil==5.9.6  # System monitoring

# Development and testing (optional)
# pytest==7.4.3
# pytest-flask==1.3.0
